Started to addd buttons
Started to add the buttons and teh vectors to hold the randomized
characters but encounter issue with button::widget defiinition.

Widget::Button only takes a const string& as an argument and i need to
pass it a non-const char&??????

Will talk to TA's and Prof about if i can add more funcitonality to teh
Button definition in our graphics library.